+++ src/main/java/io/hhplus/concert/domain/support/DistributedSpinLock.java
@@ -1,4 +1,4 @@
-package io.hhplus.concert.infrastructure.distributedlocks;
+package io.hhplus.concert.domain.support;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,8 +7,10 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface DistributedSimpleLock {
+public @interface DistributedSpinLock {
 	String key();
 	int ttlSeconds() default 5;
+	int waitSeconds() default 5;
+	int retryMillis() default 100;
 	String prefix() default "lock";
 }
