+++ src/test/java/io/hhplus/concert/ConcertApplicationTests.java
@@ -3,7 +3,8 @@ package io.hhplus.concert;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+import io.hhplus.concert.infrastructure.containers.TestcontainersConfiguration;
 
 @Import(TestcontainersConfiguration.class)
 @SpringBootTest
