+++ src/test/java/io/hhplus/concert/application/usecase/payment/PaymentUsecaseIntegrationTest.java
@@ -22,7 +22,7 @@ import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.jdbc.Sql;
 
-import io.hhplus.concert.TestcontainersConfiguration;
+import io.hhplus.concert.infrastructure.containers.TestcontainersConfiguration;
 import io.hhplus.concert.domain.concert.Concert;
 import io.hhplus.concert.domain.concert.ConcertDate;
 import io.hhplus.concert.domain.concert.ConcertDateRepository;
@@ -33,7 +33,6 @@ import io.hhplus.concert.domain.concert.ConcertService;
 import io.hhplus.concert.domain.payment.Payment;
 import io.hhplus.concert.domain.payment.PaymentRepository;
 import io.hhplus.concert.domain.payment.PaymentService;
-import io.hhplus.concert.domain.payment.PaymentServiceIntegrationTest;
 import io.hhplus.concert.domain.reservation.Reservation;
 import io.hhplus.concert.domain.reservation.ReservationCommand;
 import io.hhplus.concert.domain.reservation.ReservationInfo;
