+++ src/test/java/io/hhplus/concert/infrastructure/containers/TestcontainersConfiguration.java
@@ -1,17 +1,10 @@
-package io.hhplus.concert;
-
-import java.time.Duration;
-
-import jakarta.annotation.PreDestroy;
+package io.hhplus.concert.infrastructure.containers;
 
 import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 import org.testcontainers.containers.MySQLContainer;
-import org.testcontainers.containers.wait.strategy.Wait;
 import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
 
 @TestConfiguration
