+++ src/test/java/io/hhplus/concert/interfaces/api/UserE2EIntegrationTest.java
@@ -21,22 +21,17 @@ import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.MediaType;
-import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.jdbc.Sql;
-import org.springframework.transaction.annotation.Transactional;
 
-import io.hhplus.concert.TestcontainersConfiguration;
+import io.hhplus.concert.infrastructure.containers.TestcontainersConfiguration;
 import io.hhplus.concert.domain.token.Token;
-import io.hhplus.concert.domain.token.TokenInfo;
 import io.hhplus.concert.domain.token.TokenRepository;
 import io.hhplus.concert.domain.token.TokenService;
 import io.hhplus.concert.domain.user.User;
 import io.hhplus.concert.domain.user.UserCommand;
 import io.hhplus.concert.domain.user.UserInfo;
-import io.hhplus.concert.domain.user.UserPoint;
 import io.hhplus.concert.domain.user.UserPointRepository;
 import io.hhplus.concert.domain.user.UserRepository;
 import io.hhplus.concert.domain.user.UserService;
